pyproject.toml
[tool.poetry]
name = "drf-project-template"
version = "0.1.0"
description = ""
authors = ["nerdthejohn <newtonjohn043@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_project_template"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.3"
django-rest-framework = "^0.1.0"
celery = "^5.3.1"
redis = "^4.6.0"
psycopg2-binary = "^2.9.6"
python-decouple = "^3.8"
django-celery-beat = "^2.5.0"
black = "^23.3.0"
drf-spectacular = "^0.26.3"
dj-rest-auth = "^4.0.1"
django-allauth = "^0.54.0"
pyotp = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
